working on react tests
